+++ libs/frontend/abstract/application/src/renderer/runtime-component/runtime-component.model.interface.ts
@@ -31,7 +31,6 @@ export interface IRuntimeComponentModel extends AnyModel {
   /**
    * Exposed for external use by other models and to preserve structure
    */
-  runtimeElementsList: Array<IRuntimeElementModel>
   runtimeParent?: Ref<IRuntimeElementModel>
   runtimeProps: IRuntimeComponentPropModel
   runtimeRootElement: IRuntimeElementModel
